expanded monster dictionary, paginated profile messages
added more monsters, tried to lend more variance to elemental types this time.

reduced the elemental crits from x2 damage to x1.5. the final number is rounded.

tidied up the look and feel of the //me command. paginated bot messages are now sent to prevent massive messages, such as if the user has a lot of items.